Add Factorys for dependency inversion

+++ ProjectJavaOOP/src/terminal/displayed/ResultDisplayedFactory.java
@@ -6,9 +6,9 @@ import terminal.executable.executables.CommandExecutable;
 public class ResultDisplayedFactory {
 
     public ResultDisplayed create(Command command, CommandExecutable commandExecutable) {
-        if (command.isAddCommand()) {
+        if (command.isAddCommand() && command.isStudentCommand()) {
             return new CreateStudentDisplayed(commandExecutable);
-        } else if (command.isDeleteCommand()) {
+        } else if (command.isDeleteCommand() && command.isStudentCommand()) {
             return new DeleteStudentDisplayed(commandExecutable);
         } else return new NoneResultDisplayed();
     }
